Add new target "freepastry-core-jar", which creates a "pastry-core.jar" file excluding Glacier, SplitStream, and rice.pastry.testing.*.  Like freepastry-jar, this target depends on beautifythencompile.
$ ls -l pastry*jar
-rw-r--r--   1 dsandler  dsandler   785771 Aug 10 01:00 pastry-core.jar
-rw-r--r--   1 dsandler  dsandler  1167450 Aug 10 01:06 pastry.jar

(Approx 372 K saved in jar as of this checkin.)

This checkin also normalizes the names of common targets to use hyphens and
not underscores; ergo, freepastry_jar -> freepastry-jar, epost_compile ->
epost-compile, etc.  (Most of the other newer targets used hyphens so this
seemed appropriate.)  If you have cron jobs that use these old target names
(instead of just `ant` or `ant jar`) please be sure to update to the new
target.